Implement Client#wait_for_result method.
This method performs looping over the result-retrieve operation until
the result has been acknowledged. This is important for some if not most
operations, though currently only a few methods will make use of it.

Additionally there is both a class and instance method of
:set_result_poll_interval to change this from the default of 0.5
seconds to a user defined value.